+++ primefaces-integration-tests/src/test/java/org/primefaces/integrationtests/confirmdialog/ConfirmDialog002Test.java
@@ -47,7 +47,7 @@ class ConfirmDialog002Test extends AbstractPrimePageTest {
     void showDialog(Page page) {
         // Arrange
         ConfirmDialog dialog = page.dialog;
-        noMinLoadAnim();
+        noAjaxMinLoadAnimation();
         assertFalse(dialog.isVisible());
 
         // Act
@@ -65,7 +65,7 @@ class ConfirmDialog002Test extends AbstractPrimePageTest {
     void showWidget(Page page) {
         // Arrange
         ConfirmDialog dialog = page.dialog;
-        noMinLoadAnim();
+        noAjaxMinLoadAnimation();
 
         // Act
         dialog.show();
@@ -80,7 +80,7 @@ class ConfirmDialog002Test extends AbstractPrimePageTest {
     void hideWidget(Page page) {
         // Arrange
         ConfirmDialog dialog = page.dialog;
-        noMinLoadAnim();
+        noAjaxMinLoadAnimation();
         dialog.show();
         assertTrue(dialog.isVisible());
 
@@ -97,7 +97,7 @@ class ConfirmDialog002Test extends AbstractPrimePageTest {
     void confirmNo(Page page) {
         // Arrange
         ConfirmDialog dialog = page.dialog;
-        noMinLoadAnim();
+        noAjaxMinLoadAnimation();
         assertFalse(dialog.isVisible());
         page.confirm.click();
 
@@ -115,7 +115,7 @@ class ConfirmDialog002Test extends AbstractPrimePageTest {
     void confirmYes(Page page) {
         // Arrange
         ConfirmDialog dialog = page.dialog;
-        noMinLoadAnim();
+        noAjaxMinLoadAnimation();
         page.confirm.click();
         assertTrue(dialog.isVisible());
 
@@ -133,7 +133,7 @@ class ConfirmDialog002Test extends AbstractPrimePageTest {
     void deleteNo(Page page) {
         // Arrange
         ConfirmDialog dialog = page.dialog;
-        noMinLoadAnim();
+        noAjaxMinLoadAnimation();
         assertFalse(dialog.isVisible());
         page.delete.click();
 
@@ -151,7 +151,7 @@ class ConfirmDialog002Test extends AbstractPrimePageTest {
     void deleteYes(Page page) {
         // Arrange
         ConfirmDialog dialog = page.dialog;
-        noMinLoadAnim();
+        noAjaxMinLoadAnimation();
         assertFalse(dialog.isVisible());
         page.delete.click();