+++ primefaces/src/main/resources/META-INF/resources/primefaces/forms/forms.commandlink.js
@@ -68,7 +68,7 @@ PrimeFaces.widget.CommandLink = PrimeFaces.widget.BaseWidget.extend({
                     }
                     PrimeFaces.queueTask(
                         function(){ $this.endAjaxDisabled($this); },
-                        Math.max(PrimeFaces.ajax.minLoadAnim + $this.ajaxStart - Date.now(), 0)
+                        Math.max(PrimeFaces.ajax.minLoadAnimation + $this.ajaxStart - Date.now(), 0)
                     );
                     delete $this.ajaxStart;
                 }