+++ docs/14_0_0/gettingstarted/whatsnew.md
@@ -11,7 +11,7 @@ Look into [migration guide](https://primefaces.github.io/primefaces/14_0_0/#/../
 * Core
     * `MOVE_SCRIPTS_TO_BOTTOM` adds new option `defer` to defer loading scripts.
     * OS settings for `prefers-reduced-motion: reduce` is now respected and PF disables all animations.
-    * Minimum inline Ajax load animation duration via `PrimeFaces.ajax.minLoadAnim`.
+    * Minimum inline Ajax load animation duration via `PrimeFaces.ajax.minLoadAnimation`.
 
 * ConfirmDialog
    * Added `ajax=false` support.